+++ src/BiotSavart/ka_utils.jl
@@ -92,9 +92,15 @@ struct PseudoGPU <: KA.GPU end
 KA.isgpu(::PseudoGPU) = false  # needed to be considered as a CPU backend by KA
 KA.allocate(::PseudoGPU, ::Type{T}, dims::Dims) where {T} = KA.allocate(KA.CPU(), T, dims)
 KA.synchronize(::PseudoGPU) = nothing
-KA.ndevices(::PseudoGPU) = 1
-KA.device(::PseudoGPU) = 1
-KA.device!(::PseudoGPU, device::Int) = device == 1 ? nothing : throw(ArgumentError("device must be 1"))
+
+# Emulate multiple devices using task-local storage
+KA.ndevices(::PseudoGPU) = 4
+KA.device(::PseudoGPU) = get(task_local_storage(), :PseudoGPU_device, 1)::Int
+function KA.device!(backend::PseudoGPU, device::Int)
+    1 ≤ device ≤ KA.ndevices(backend) || throw(ArgumentError("invalid device"))
+    task_local_storage(:PseudoGPU_device, device)
+end
+
 KA.copyto!(::PseudoGPU, u, v) = copyto!(u, v)
 Adapt.adapt(::PseudoGPU, u::Array) = copy(u)  # simulate host → device copy (making sure arrays are not aliased)