+++ CHANGELOG.md
@@ -7,10 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.32.1] - 2025-11-22
 
+### Changed
+
+- **[Slightly breaking]**: In `ParamsBiotSavart`, `quadrature_near_singularity` can no
+  longer be an `AdaptiveTanhSinh` quadrature, and must be a statically-sized quadrature
+  (typically `GaussLegendre(n)`).
+
 ### Added
 
 - Reimplement self-interaction Biot–Savart term on GPU. It is also much faster than before
   on CPU.
+
+- Reimplement local Biot–Savart term (LIA) on GPU.
+
+- `PointData` now includes curve derivatives on nodes. This is useful for the local term,
+  but can also be useful for diagnostics or forcing.
 
 ## [0.32.0] - 2025-11-21